Add kpi merging to articles
